requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
hvac==2.0.0
requests==2.31.0
cryptography==41.0.7
PyMySQL==1.1.0
SQLAlchemy==2.0.23
alembic==1.13.1
gunicorn==21.2.0
